+++ tests/02_eradiate/02_system/test_basic.py
@@ -66,8 +66,8 @@ def test_radiometric_accuracy(modes_all_mono, illumination, spp, li, ert_seed_st
     with pytest.warns(UserWarning, match="the selected mode is single-precision") if (
         eradiate.mode().is_single_precision and spp > 100000
     ) else nullcontext():
-        measure = MultiDistantMeasure.from_viewing_angles(
-            zeniths=vza, azimuths=0.0, target=[0, 0, 0], spp=spp
+        measure = MultiDistantMeasure.hplane(
+            zeniths=vza, azimuth=0.0, target=[0, 0, 0], spp=spp
         )
 
     objects = {