+++ tests/02_eradiate/03_regression/atmospheres/test_rpv_afgl1986.py
@@ -49,8 +49,8 @@ def test_rpv_afgl1986_brfpp(mode_ckd_double, artefact_dir, session_timestamp):
         measures=[
             {
                 "type": "mdistant",
-                "construct": "from_viewing_angles",
-                "azimuths": 0.0,
+                "construct": "hplane",
+                "azimuth": 0.0,
                 "zeniths": np.arange(-75.0, 75.01, 2.0),
                 "srf": {"type": "multi_delta", "wavelengths": 550 * ureg.nm},
             }