+++ tests/02_eradiate/03_regression/romc/test_het04.py
@@ -99,9 +99,9 @@ def test_het04a1_brfpp(mode_mono_double, artefact_dir, session_timestamp):
         measures=[
             {
                 "type": "mdistant",
-                "construct": "from_viewing_angles",
+                "construct": "hplane",
                 "spp": 20000,
-                "azimuths": [180] * ureg.deg,
+                "azimuth": 180 * ureg.deg,
                 "zeniths": np.arange(-75, 75.01, 2) * ureg.deg,
                 "target": {
                     "type": "rectangle",